minor fix , refactor

+++ Guppy.java
@@ -180,8 +180,8 @@ public class Guppy extends Fish {
 
             //double atan2(double y, double x)
             double a = Math.atan2(F.get(idx).get_y()-get_y(), F.get(idx).get_x()-get_x());
-            set_x(get_x()+(double)(get_speed()*Math.cos(a)*1.5));
-            set_y(get_y()+(double)(get_speed()*Math.sin(a)*1.5));
+            set_x(get_x() + (get_speed() * Math.cos(a) * 1.5));
+            set_y(get_y() + (get_speed() * Math.sin(a) * 1.5));
         }
         else {
             //random arahnya
